Cleanup large result set panel
